Finished the file that creates the big database
Now we can use this file to create a database in our convenient time.
We can run a subsequence of songs each time and save it as the current database. In addition I created a log file so we can track the process.

Note: the big_one_token_dataset.csv is rewritten every 10 songs. This will be a problem if we have a huge dataset. But, we are working with short strings so it thought it will do the job in our case.