Combined model progress
Figured out how to tokenize the target Hebrew sentence the same as the last translator. Now we need to figure out how to calculate the loss but its should be an easier task with this new understanding.
In addition, I have added this to the combined model but its still doesn't work well. Still need to fix the loss calculation. Working on it now.