First model that minimizes the loss!
I wanted to check if our implementation is good so I punished the model only on the generated tokens and it works beautifully. Its a sign that our current implementation is pretty good for a start and we need to focus on improving the transformers alone before we start training the combined model with them. As it is right now they don't work well enough and the combined model doesn't seem to learn with them.