Current Progress - Combined model is working nicely with minimal training
It looks like the combined model is working fine. I trained it on the first 2 tokens in each sentence (750 sentences) with the pretrained model of transformer 1 and a random weighted transformer 2 and tested it on words he had seen and it outputs nice results for so little training. Now we want to train it with a lot more data and see what happens.