Test Results So Far
I have tested most of the models (Before & After training) and saved the results so far.
Our model works better after training with 26 Successes / 5K. Its not much but its a start! Specially because the LLM as is succeeded only 6/5k.
But we have work to do cuase after fine tune the LLM succeeded in 296/5k.